er-ccapi-python/pyproject.toml
[tool.poetry]
name = "er-ccapi-python"
version = "0.1.0"
description = ""
authors = ["Energy Robotics GmbH"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
get-docker-secret = "^2.0.0"
requests = "^2.31.0"
apscheduler = "^3.10.4"
rich = "^13.6.0"
pyyaml = "^6.0.1"
gql = {extras = ["all"], version = "^3.4.1"}
pandas = "^2.1.3"



[tool.poetry.group.dev.dependencies]
mkdocs-material = "^9.5.2"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120